requirements.txt
pinecone
langchain-pinecone
langchain-openai
langchain-text-splitters
langchain
python-dotenv
pypdf
boto3
langchain-aws
streamlit
ai21